# services/proj_wesbank/backend/src/app/api/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Wesbank API",
    "description": "Your Site API ⚡",
    "termsOfService": "https://www.YourSite.co.za/termsandconditions.html",
    "contact": { "url": "https://www.YourSite.co.za/#contact" },
    "version": "1.5.0"
  },
  "paths": {
...
    "/v0/sho002_spend_per_month_per_component_graph": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Spend Per Month Per Component",
        "description": "Get spend per month",
        "operationId": "sho002_spend_per_month_per_component_v0_sho002_spend_per_month_per_component_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_least_assets_costs_per_component": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Least Assets Costs Per Component",
        "description": "Get the bottom vehicles cost per component",
        "operationId": "sho002_least_assets_costs_per_component_v0_sho002_least_assets_costs_per_component_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_total_counts_and_costs_for_completed_accrual_orderExcep_and_invoiceExcep": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Get Total Counts And Costs For Completed Accrual Orderexcept And Invoiceexcept",
        "description": "get total counts and costs for completed accrual, order exceptions and invoice exceptions",
        "operationId": "sho002_get_total_counts_and_costs_for_completed_accrual_orderExcept_and_invoiceExcept_v0_sho002_get_total_counts_and_costs_for_completed_accrual_orderExcep_and_invoiceExcep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_inv_by_component": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Get Inv By Component",
        "description": "get inv by component",
        "operationId": "sho002_get_inv_by_component_v0_sho002_get_inv_by_component_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_maintenance_cost_over_odo": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Get Maintenance Cost Over Odo",
        "description": "get maintenance maps",
        "operationId": "sho002_get_maintenance_cost_over_odo_v0_sho002_get_maintenance_cost_over_odo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/v0/sho002_best_assets_costs_per_component": {
      "post": {
        "tags": ["components"],
        "summary": "Sho002 Best Assets Costs Per Component",
        "description": "Get the bottom vehicles cost per component",
        "operationId": "sho002_best_assets_costs_per_component_v0_sho002_best_assets_costs_per_component_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "token" } }
      }
    }
  }
}